requirements.txt
click==8.0.3
creme==0.6.1
dnspython==2.2.0
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
mmh3==2.5.1
nltk==3.6.7
numpy==1.19.5
pandas==1.4.0
passlib==1.7.4
pymongo==4.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2022.1.18
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
tqdm==4.62.3
Werkzeug==2.0.2
